Rewrite tests for Posting model
- TransactionServiceTests assert on result.Postings (AccountId/Amount) instead of removed SourceAccountId/DestinationAccountId
- FinancialAnalyticsServiceTests drop Transaction preset, use local categoryId/sum; decimal money types
- SpendingCategoryModel.TotalSpending -> decimal

+++ Service.Models/Models/FinancialAnalytics/SpendingCategoryModel.cs
@@ -5,4 +5,4 @@
 /// </summary>
 /// <param name="TotalSpending">Общая сумма расходов</param>
 /// <param name="Categories">Категории с суммами трат</param>
-public sealed record SpendingCategoryModel(double TotalSpending, IReadOnlyCollection<CategorySpendingItemModel> Categories);
+public sealed record SpendingCategoryModel(decimal TotalSpending, IReadOnlyCollection<CategorySpendingItemModel> Categories);
